# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login to system",
                "operationId": "login",
                "requestBody": {
                    "description": "Login to System",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "deprecated": false
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Per Page Product Count",
                "operationId": "6d8ae74dcafe19c4813904d4ea9e4c57",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Order By Column Name",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order ASC or DESC",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "description": "Create product",
                "operationId": "9f0b814cd8d69075ed70d4cc79c9f417",
                "requestBody": {
                    "description": "Create product",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string",
                                        "example": "Oshxona mebeli"
                                    },
                                    "slug": {
                                        "description": "Product slug",
                                        "type": "string",
                                        "example": "product-title"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "integer",
                                        "example": "682"
                                    },
                                    "image": {
                                        "description": "Product image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Per Page Product Count",
                "operationId": "f6410f1810b0a90ba3d72d5b8250a27d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product Not Found"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product",
                "operationId": "ae6d19da91d293e4d27eeb1a67b836ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "description": "method",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update product",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "slug"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string",
                                        "example": "Oshxona mebeli"
                                    },
                                    "slug": {
                                        "description": "Product slug",
                                        "type": "string",
                                        "example": "product-title"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "integer",
                                        "example": "682"
                                    },
                                    "image": {
                                        "description": "Product image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "description": "Delete specific product",
                "operationId": "3dda9543e42196cc1ff7b18d2c9f9357",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Product Not Found"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User profile",
                "description": "User profile",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "description": "User logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register to system",
                "operationId": "Register",
                "requestBody": {
                    "description": "Register to System",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "Muzaffar Doe"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "muzaffar@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "deprecated": false
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}